moved NetPrintsEditorFactory to seperate file, removed unused interfaces, add VS project references to netprints project
